requirements.txt
Flask==2.2.2
gunicorn==20.1.0
python-decouple==3.6
python-dotenv==0.21.0
requests==2.28.1
Flask-Cors==3.0.10

